move auth.ts from src/app/auth.ts -> src/auth.ts
